+++ platform/dashboard/src/main.tsx
@@ -4,6 +4,7 @@ import 'virtual:svg-icons-register'
 
 import { App } from './App'
 import './index.css'
+import { HelmetProvider } from 'react-helmet-async'
 
 if (import.meta.env.NODE_ENV === 'development') {
   import('~/lib/logger')
@@ -11,6 +12,8 @@ if (import.meta.env.NODE_ENV === 'development') {
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
-    <App />
+    <HelmetProvider>
+      <App />
+    </HelmetProvider>
   </React.StrictMode>
 )